organizing WalletAssetsScreen
