+++ src/screens/WalletAssetsScreen/AssetList/Card/types.ts
@@ -1,0 +1,10 @@
+import { TokenType } from '@styles';
+
+export interface CardProps {
+	onPress: () => void;
+	coinName: string;
+	coinSymbol: TokenType;
+	walletBalance: number;
+	walletBalanceUsd: number;
+	interest?: string;
+}
