+++ src/screens/WalletAssetsScreen/AssetList/Header/index.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import React from 'react';
 import { Text } from '@components';
 import { View } from 'react-native';
