+++ src/screens/WalletAssetsScreen/AssetList/Selector/Button/styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 0,
+		paddingHorizontal: 12,
+		height: '100%',
+		borderRadius: 24,
+		borderWidth: 1,
+		alignItems: 'center',
+		justifyContent: 'center',
+		alignSelf: 'center',
+		marginRight: 8
+	}
+});
