+++ src/screens/WalletAssetsScreen/AssetList/Selector/Button/types.ts
@@ -1,0 +1,6 @@
+interface ButtonProps {
+	onPress: () => void;
+	active: boolean;
+}
+
+export type { ButtonProps };
