+++ src/screens/WalletAssetsScreen/AssetList/Selector/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		height: 31,
+		paddingHorizontal: 24,
+		flexDirection: 'row'
+	}
+});
