+++ src/screens/WalletAssetsScreen/AssetList/Selector/types.tsx
@@ -1,0 +1,4 @@
+export interface AssetSelectorProps {
+	active: number;
+	setActive: (active: number) => void;
+}
