--- src/screens/WalletAssetsScreen/AssetList/coins.json
@@ -1,252 +1,0 @@
-[
-	{
-		"id": "bitcoin",
-		"name": "Bitcoin",
-		"symbol": "btc"
-	},
-	{
-		"id": "ethereum",
-		"name": "Ethereum",
-		"symbol": "eth"
-	},
-	{
-		"id": "tether",
-		"name": "Tether",
-		"symbol": "usdt"
-	},
-	{
-		"id": "binancecoin",
-		"name": "BNB",
-		"symbol": "bnb"
-	},
-	{
-		"id": "usd-coin",
-		"name": "USD Coin",
-		"symbol": "usdc"
-	},
-	{
-		"id": "ripple",
-		"name": "XRP",
-		"symbol": "xrp"
-	},
-	{
-		"id": "cardano",
-		"name": "Cardano",
-		"symbol": "ada"
-	},
-	{
-		"id": "solana",
-		"name": "Solana",
-		"symbol": "sol"
-	},
-	{
-		"id": "avalanche-2",
-		"name": "Avalanche",
-		"symbol": "AVAX"
-	},
-	{
-		"id": "terra-luna",
-		"name": "Terra",
-		"symbol": "luna"
-	},
-	{
-		"id": "polkadot",
-		"name": "Polkadot",
-		"symbol": "dot"
-	},
-	{
-		"id": "dogecoin",
-		"name": "Dogecoin",
-		"symbol": "doge"
-	},
-	{
-		"id": "binance-usd",
-		"name": "Binance USD",
-		"symbol": "busd"
-	},
-	{
-		"id": "shiba-inu",
-		"name": "Shiba Inu",
-		"symbol": "shib"
-	},
-	{
-		"id": "terrausd",
-		"name": "TerraUSD",
-		"symbol": "ust"
-	},
-	{
-		"id": "matic-network",
-		"name": "Polygon",
-		"symbol": "matic"
-	},
-	{
-		"id": "crypto-com-chain",
-		"name": "Cronos",
-		"symbol": "cro"
-	},
-	{
-		"id": "wrapped-bitcoin",
-		"name": "Wrapped Bitcoin",
-		"symbol": "wbtc"
-	},
-	{
-		"id": "dai",
-		"name": "Dai",
-		"symbol": "dai"
-	},
-	{
-		"id": "litecoin",
-		"name": "Litecoin",
-		"symbol": "ltc"
-	},
-	{
-		"id": "cosmos",
-		"name": "Cosmos",
-		"symbol": "atom"
-	},
-	{
-		"id": "chainlink",
-		"name": "Chainlink",
-		"symbol": "link"
-	},
-	{
-		"id": "tron",
-		"name": "TRON",
-		"symbol": "trx"
-	},
-	{
-		"id": "near",
-		"name": "Near",
-		"symbol": "near"
-	},
-	{
-		"id": "bitcoin-cash",
-		"name": "Bitcoin Cash",
-		"symbol": "bch"
-	},
-	{
-		"id": "algorand",
-		"name": "Algorand",
-		"symbol": "algo"
-	},
-	{
-		"id": "ftx-token",
-		"name": "FTX Token",
-		"symbol": "ftt"
-	},
-	{
-		"id": "leo-token",
-		"name": "LEO Token",
-		"symbol": "leo"
-	},
-	{
-		"id": "staked-ether",
-		"name": "Lido Staked Ether",
-		"symbol": "steth"
-	},
-	{
-		"id": "okb",
-		"name": "OKB",
-		"symbol": "okb"
-	},
-	{
-		"id": "stellar",
-		"name": "Stellar",
-		"symbol": "xlm"
-	},
-	{
-		"id": "fantom",
-		"name": "Fantom",
-		"symbol": "ftm"
-	},
-	{
-		"id": "uniswap",
-		"name": "Uniswap",
-		"symbol": "uni"
-	},
-	{
-		"id": "hedera-hashgraph",
-		"name": "Hedera",
-		"symbol": "hbar"
-	},
-	{
-		"id": "decentraland",
-		"name": "Decentraland",
-		"symbol": "mana"
-	},
-	{
-		"id": "internet-computer",
-		"name": "Internet Computer",
-		"symbol": "icp"
-	},
-	{
-		"id": "axie-infinity",
-		"name": "Axie Infinity",
-		"symbol": "axs"
-	},
-	{
-		"id": "ethereum-classic",
-		"name": "Ethereum Classic",
-		"symbol": "etc"
-	},
-	{
-		"id": "the-sandbox",
-		"name": "The Sandbox",
-		"symbol": "sand"
-	},
-	{
-		"id": "elrond-erd-2",
-		"name": "Elrond",
-		"symbol": "egld"
-	},
-	{
-		"id": "vechain",
-		"name": "VeChain",
-		"symbol": "vet"
-	},
-	{
-		"id": "filecoin",
-		"name": "Filecoin",
-		"symbol": "fil"
-	},
-	{
-		"id": "klay-token",
-		"name": "Klaytn",
-		"symbol": "klay"
-	},
-	{
-		"id": "theta-token",
-		"name": "Theta Network",
-		"symbol": "theta"
-	},
-	{
-		"id": "tezos",
-		"name": "Tezos",
-		"symbol": "xtz"
-	},
-	{
-		"id": "ecomi",
-		"name": "ECOMI",
-		"symbol": "omi"
-	},
-	{
-		"id": "monero",
-		"name": "Monero",
-		"symbol": "xmr"
-	},
-	{
-		"id": "cdai",
-		"name": "cDAI",
-		"symbol": "cdai"
-	},
-	{
-		"id": "compound-ether",
-		"name": "cETH",
-		"symbol": "ceth"
-	},
-	{
-		"id": "magic-internet-money",
-		"name": "Magic Internet Money",
-		"symbol": "mim"
-	}
-]
