+++ src/screens/WalletAssetsScreen/AssetList/styles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		paddingTop: 32,
+		height: '100%'
+	}
+});
