+++ src/screens/WalletAssetsScreen/AssetList/types.ts
@@ -1,0 +1,5 @@
+import { WalletToken } from '@models/wallet';
+
+export interface AssetListProps {
+	walletTokens: WalletToken[];
+}
